requirements.txt
torch
torchvision
numpy
matplotlib
scikit-image
pillow
scikit-learn
tensorboard
flask
streamlit
tqdm
